Update based on Copilot feedback
Changes in this commit

- Removed aria labels as recommended by Copilot.
- Change type `any` to `WeatherForecast` as recommended by Copilot. This required an update to the `Import` statement.
- Changed weatherforecast url to be relative as recommended by Copilot.
- Changed the constructor to use `ngOnInit` instead.

I noticed that the data wasn't reliably showing up on the page. I had to add a change detector to get it to reliably appear.
I'm not an Angular expert, so there may be a way to remove the change detector, but I wasn't able to figure it out.

I think we can remove `@Injectable()` but I didn't want to add that to this PR.